m68k: SYSCALL_DEFINEx

+++ arch/m68k/include/asm/syscalls.h
@@ -7,11 +7,11 @@
 
 asmlinkage long sys_m68k_cacheflush(unsigned long addr, int scope, int cache,
 			      unsigned long len);
-asmlinkage int sys_atomic_cmpxchg_32(unsigned long newval, int oldval, int d3,
+asmlinkage long sys_atomic_cmpxchg_32(unsigned long newval, int oldval, int d3,
 				     int d4, int d5, unsigned long __user *mem);
-asmlinkage unsigned long sys_get_thread_area(void);
-asmlinkage int sys_set_thread_area(unsigned long tp);
-asmlinkage int sys_atomic_barrier(void);
+asmlinkage long sys_get_thread_area(void);
+asmlinkage long sys_set_thread_area(unsigned long tp);
+asmlinkage long sys_atomic_barrier(void);
 
 #include <asm-generic/syscalls.h>
